% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite-name.R
\name{as.sqlite_name}
\alias{as.sqlite_name}
\alias{as.character.sqlite_name}
\alias{as.sqlite_name.character}
\title{Coerce to a SQLite Name}
\usage{
as.sqlite_name(x, ...)

\method{as.character}{sqlite_name}(x, ...)

\method{as.sqlite_name}{character}(x, ...)
}
\arguments{
\item{x}{The object to coerce}

\item{...}{Unused.}
}
\description{
Coerce to a SQLite Name
}
\section{Methods (by class)}{
\itemize{
\item \code{sqlite_name}: Coerces SQLite name vector to a character vector

\item \code{character}: Coerces character vector to a SQLite name vector
}}

\examples{
as.sqlite_name("Data")
}
