% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_sqlite_name}
\alias{check_sqlite_name}
\title{Check SQLite Name}
\usage{
check_sqlite_name(x, values = NULL, length = NA, unique = FALSE,
  sorted = FALSE, named = NA, attributes = named,
  x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{values}{NULL or a vector specifying the values.}

\item{length}{A flag indicating whether x should have elements (versus no elements) or a missing value indicating no requirements or a count or count range of the number of elements or a count vector of the permitted number of elements.}

\item{unique}{A flag indicating whether the values must be unique.}

\item{sorted}{A flag indicating whether the vector must be sorted.}

\item{named}{A flag indicating whether the vector must be named or unnamed or NA if it doesn't matter.}

\item{attributes}{A flag indicating whether the vector must or must not have attributes or NA if it doesn't matter.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of the original object.
}
\description{
Check SQLite Name
}
\examples{
check_sqlite_name("Data")
}
