% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R, R/deprecated.R
\name{chk_sqlite_conn}
\alias{chk_sqlite_conn}
\alias{check_sqlite_connection}
\title{Check SQLite Connection}
\usage{
chk_sqlite_conn(x, connected = NA, err = TRUE, x_name = NULL)

check_sqlite_connection(x, connected = NA, x_name = substitute(x),
  error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{connected}{A logical scalar specifying whether x should be connected.}

\item{err}{A flag indicating whether to throw an informative error 
or immediately generate an informative message if the check fails.}

\item{x_name}{A string of the name of object x or NULL.}

\item{error}{A flag indicating whether to throw an informative error 
or immediately generate an informative message if the check fails.}
}
\value{
TRUE if passes check. Otherwise if throws an informative error unless
\code{err = FALSE} in which case it returns FALSE.
}
\description{
Checks whether an R object is a SQLite Connection.
}
\section{Functions}{
\itemize{
\item \code{check_sqlite_connection}: Check SQLite Connection
}}

\examples{
conn <- rws_connect()
chk_sqlite_conn(conn)
rws_disconnect(conn)
try(chk_sqlite_conn(conn, connected = TRUE))
}
