% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{dbReadLogTableSQLite}
\alias{dbReadLogTableSQLite}
\title{Read dbWriteSQLiteLog Table}
\usage{
dbReadLogTableSQLite(conn = getOption("dbWriteSQLite.conn", NULL))
}
\arguments{
\item{conn}{A \code{\linkS4class{SQLiteConnection}} object.}
}
\value{
A data frame of the dbWriteSQLiteLog table
}
\description{
The table is created if it doesn't exist.
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite())
dbReadLogTableSQLite(con)
DBI::dbDisconnect(con)
}
