% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{dbReadMetaTableSQLite}
\alias{dbReadMetaTableSQLite}
\title{Read dbWriteSQLiteMeta Table}
\usage{
dbReadMetaTableSQLite(conn = getOption("dbWriteSQLite.conn", NULL))
}
\arguments{
\item{conn}{A \code{\linkS4class{SQLiteConnection}} object.}
}
\value{
A data frame of the dbWriteSQLiteMeta table
}
\description{
The table is created if it doesn't exist.
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite())
dbReadMetaTableSQLite(con)
DBI::dbDisconnect(con)
}
