% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{dbReadTableSQLite}
\alias{dbReadTableSQLite}
\title{Read a Table from a SQLite Database}
\usage{
dbReadTableSQLite(table_name, conn = getOption("dbWriteSQLite.conn",
  NULL), meta = TRUE)
}
\arguments{
\item{table_name}{A string of the (case insensitive) table name.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} object.}

\item{meta}{A flag specifying whether to preserve meta data.}
}
\value{
A data frame (tibble if tibble package is installed) of the table.
}
\description{
Read a Table from a SQLite Database
}
\seealso{
Other dbReadTableSQLite: \code{\link{dbReadTablesSQLite}}
}
\concept{dbReadTableSQLite}
