% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-table.R
\name{dbRemoveTableSQLite}
\alias{dbRemoveTableSQLite}
\title{Write a data frame to a SQLite Database}
\usage{
dbRemoveTableSQLite(conn, name, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{SQLiteConnection}} object, produced by
[DBI::dbConnect()].}

\item{name}{A string of the (case insensitive) table name.}

\item{...}{Errors if used.}
}
\description{
Write a data frame to a SQLite Database
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite())
DBI::dbDisconnect(con)
}
