% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-table.R
\docType{methods}
\name{dbWriteTable,SQLiteConnection,character,data.frame-method}
\alias{dbWriteTable,SQLiteConnection,character,data.frame-method}
\title{Write a local data frame or file to the database}
\usage{
\S4method{dbWriteTable}{SQLiteConnection,character,data.frame}(conn, name,
  value, ...,
  row.names = pkgconfig::get_config("RSQLite::row.names.table", FALSE),
  overwrite = FALSE, append = FALSE, field.types = NULL,
  temporary = FALSE)
}
\arguments{
\item{conn}{a \code{\linkS4class{SQLiteConnection}} object, produced by
[DBI::dbConnect()]}

\item{name}{a character string specifying a table name. SQLite table names
are \emph{not} case sensitive, e.g., table names `ABC` and `abc`
are considered equal.}

\item{...}{Needed for compatibility with generic. Otherwise ignored.}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

If \code{TRUE}, always translate row names to a column called "row_names".
If \code{FALSE}, never translate row names. If \code{NA}, translate
rownames only if they're a character vector.

A string is equivalent to \code{TRUE}, but allows you to override the
default name.

For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{check.names}{If `TRUE`, the default, column names will be
converted to valid R identifiers.}

\item{select.cols}{Deprecated, do not use.}
}
\description{
Functions for writing data frames or delimiter-separated files
to database tables.
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite())
DBI::dbDisconnect(con)
}
