% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{dbWriteTableSQLite}
\alias{dbWriteTableSQLite}
\title{Write a data frame to a SQLite Database}
\usage{
dbWriteTableSQLite(data, table_name = substitute(data),
  conn = getOption("dbWriteSQLite.conn", NULL), delete = FALSE,
  commit = TRUE, meta = TRUE, log = TRUE)
}
\arguments{
\item{data}{A data frame of the data to write.}

\item{table_name}{A string of the (case insensitive) table name.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} object.}

\item{delete}{A flag specifying whether to delete existing rows before inserting the data.}

\item{commit}{A flag specifying whether to commit the changes (useful for checking data).}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{log}{A flag specifying whether to log data manipulations.}
}
\value{
The updated data frame with the same columns as the table.
}
\description{
The table must exist.
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite())
DBI::dbDisconnect(con)
}
