% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{dbWriteTablesSQLite}
\alias{dbWriteTablesSQLite}
\title{Write a list of data frames to a SQLite Database}
\usage{
dbWriteTablesSQLite(list = as.list(parent.frame()),
  conn = getOption("dbWriteSQLite.conn", NULL), delete = FALSE,
  commit = TRUE, meta = TRUE, log = TRUE)
}
\arguments{
\item{list}{A named list of data frames.
Objects which are not data frames are removed.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} object.}

\item{delete}{A flag specifying whether to delete existing rows before inserting the data.}

\item{commit}{A flag specifying whether to commit the changes (useful for checking data).}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{log}{A flag specifying whether to log data manipulations.}
}
\value{
An invisible character vector of the names of the data frames that
were written to connection in the order in which they were written.
}
\description{
The tables must exist.
The commit = FALSE argument is not yet used.
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite())
DBI::dbDisconnect(con)
}
