% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite-name.R
\name{is.sqlite_name}
\alias{is.sqlite_name}
\title{Is SQLite Name}
\usage{
is.sqlite_name(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Test whether an object is a SQLite name.
}
\examples{
is.sqlite_name(as.sqlite_name("data"))
}
