% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_sqlite.character}
\alias{read_sqlite.character}
\title{Read from a SQLite Database}
\usage{
\method{read_sqlite}{character}(x,
  conn = getOption("readwritesqlite.conn", NULL), meta = TRUE, ...)
}
\arguments{
\item{x}{The object to write.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{...}{Not used.}
}
\value{
A data frame (tibble if tibble package is installed) of the table.
}
\description{
Read from a SQLite Database
}
\seealso{
Other dbReadTableSQLite: \code{\link{read_sqlite.SQLiteConnection}}
}
\concept{dbReadTableSQLite}
