% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_sqlite}
\alias{read_sqlite}
\title{Read from a SQLite Database}
\usage{
read_sqlite(x, conn = getOption("readwritesqlite.conn", NULL),
  meta = TRUE, ...)
}
\arguments{
\item{x}{An object specifying the tables to read.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{...}{Not used.}
}
\value{
A named list of data frames.
}
\description{
Read from a SQLite Database
}
\seealso{
Other write_sqlite: \code{\link{write_sqlite.data.frame}},
  \code{\link{write_sqlite.list}},
  \code{\link{write_sqlite}}
}
\concept{write_sqlite}
