% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_sqlite.SQLiteConnection}
\alias{read_sqlite.SQLiteConnection}
\title{Read from a SQLite Database}
\usage{

  \method{read_sqlite}{SQLiteConnection}(x = getOption("readwritesqlite.conn",
  NULL), meta = TRUE, ...)
}
\arguments{
\item{x}{The object to write.}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{...}{Not used.}
}
\value{
A named list of the data tables.
}
\description{
Read from a SQLite Database
}
\seealso{
Other dbReadTableSQLite: \code{\link{read_sqlite.character}}
}
\concept{dbReadTableSQLite}
