% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{read_sqlite_log}
\alias{read_sqlite_log}
\title{Read dbWriteSQLiteLog Table}
\usage{
read_sqlite_log(conn = getOption("readwritesqlite.conn", NULL))
}
\arguments{
\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}
}
\value{
A data frame of the dbWriteSQLiteLog table
}
\description{
The table is created if it doesn't exist.
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite())
read_sqlite_log(con)
DBI::dbDisconnect(con)
}
