% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{read_sqlite_meta}
\alias{read_sqlite_meta}
\title{Read Meta Data table from SQLite Database}
\usage{
read_sqlite_meta(conn = getOption("readwritesqlite.conn", NULL))
}
\arguments{
\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}
}
\value{
A data frame of the readwritesqlite_meta table
}
\description{
The table is created if it doesn't exist.
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite())
dbReadMetaTableSQLite(con)
DBI::dbDisconnect(con)
}
