% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rws_data}
\alias{rws_data}
\title{Example Data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3 rows and 7 columns.}
\usage{
rws_data
}
\description{
An sf tibble of example data.
}
\examples{
rws_data
}
\keyword{datasets}
