% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{rws_describe_meta.character}
\alias{rws_describe_meta.character}
\title{Add Descriptions to SQL Meta Data Table}
\usage{
\method{rws_describe_meta}{character}(x, column, description, ..., conn)
}
\arguments{
\item{x}{A character vector of table name(s).}

\item{column}{A character vector of column name(s).}

\item{description}{A character vector of the description(s)}

\item{...}{Not used.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}
}
\value{
An invisible character vector of the previous descriptions.
}
\description{
Add Descriptions to SQL Meta Data Table
}
\seealso{
Other rws_read_sqlite: \code{\link{rws_describe_meta.data.frame}},
  \code{\link{rws_read_sqlite.SQLiteConnection}},
  \code{\link{rws_read_sqlite.character}},
  \code{\link{rws_read_sqlite}}
}
\concept{rws_read_sqlite}
