% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{rws_describe_sqlite_meta.data.frame}
\alias{rws_describe_sqlite_meta.data.frame}
\title{Add Data Frame of Descriptions to SQL Meta Data Table}
\usage{
\method{rws_describe_sqlite_meta}{data.frame}(x, ..., conn)
}
\arguments{
\item{x}{A data frame with columns Table, Column, Description.}

\item{...}{Not used.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}
}
\value{
An invisible character vector of the previous descriptions.
}
\description{
Add Data Frame of Descriptions to SQL Meta Data Table
}
\seealso{
Other rws_read_sqlite: \code{\link{rws_describe_sqlite_meta.character}},
  \code{\link{rws_read_sqlite.SQLiteConnection}},
  \code{\link{rws_read_sqlite.character}},
  \code{\link{rws_read_sqlite}}
}
\concept{rws_read_sqlite}
