% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{rws_disconnect}
\alias{rws_disconnect}
\alias{rws_close_connection}
\title{Close SQLite Database Connection}
\usage{
rws_disconnect(conn)
}
\arguments{
\item{conn}{An objected generated by \code{\link[=SQLite]{SQLite()}}, or an existing
\code{\linkS4class{SQLiteConnection}}. If an connection, the connection
will be cloned.}
}
\description{
Closes a \code{\linkS4class{SQLiteConnection}} to a SQLite database.
}
\examples{
conn <- rws_connect()
rws_disconnect(conn)
print(conn)
}
\seealso{
\code{\link{rws_connect}()}
}
