% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{rws_export_gpkg}
\alias{rws_export_gpkg}
\title{Export all spatial datasets in a database as geopackages.}
\usage{
rws_export_gpkg(conn, dir, overwrite = FALSE)
}
\arguments{
\item{conn}{A \linkS4class{SQLiteConnection} to a database.}

\item{dir}{A string of the path to the directory to save the geopackages in.}

\item{overwrite}{A flag specifying whether to overwrite existing geopackages.}
}
\value{
An invisible named vector of the file names and new file names saved.
}
\description{
Export all spatial datasets in a database as geopackages.
}
\details{
If more than one spatial column is present in a table,
a separate geopackage will be exported for each, and the other spatial columns will be dropped.
}
