% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rws_list_tables}
\alias{rws_list_tables}
\title{Table Names}
\usage{
rws_list_tables(conn)
}
\arguments{
\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}
}
\value{
A character vector of table names.
}
\description{
Gets the table names excluding the names of the meta and log tables.
}
\examples{
conn <- rws_connect()
rws_list_tables(conn)
rws_write(rws_data, exists = FALSE, conn = conn)
rws_list_tables(conn)
rws_disconnect(conn)
}
