% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{rws_open_connection}
\alias{rws_open_connection}
\title{Opens SQLite Database Connection}
\usage{
rws_open_connection(dbname = "", exists = NA)
}
\arguments{
\item{dbname}{The path to the database file. SQLite keeps each database
instance in one single file. The name of the database \emph{is} the file
name, thus database names should be legal file names in the running
platform. There are two exceptions:

\itemize{
\item \code{""} will create a temporary on-disk database. The file
will be deleted when the connection is closed.
\item \code{":memory:"} or \code{"file::memory:"} will create a temporary
in-memory database.
}}

\item{exists}{A flag specifying whether the table(s) must already exist.}
}
\description{
Opens a \code{\linkS4class{SQLiteConnection}} to a SQLite database with
foreign key constraints enabled.
}
