% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{rws_read.character}
\alias{rws_read.character}
\title{Read Tables from a SQLite Database}
\usage{
\method{rws_read}{character}(x, meta = TRUE, conn, ...)
}
\arguments{
\item{x}{A character vector of table names.}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}

\item{...}{Not used.}
}
\value{
A named list of the data frames.
}
\description{
Read Tables from a SQLite Database
}
\seealso{
Other rws_read: \code{\link{rws_describe_meta.character}},
  \code{\link{rws_describe_meta.data.frame}},
  \code{\link{rws_read.SQLiteConnection}},
  \code{\link{rws_read}}
}
\concept{rws_read}
