% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{rws_read_init}
\alias{rws_read_init}
\alias{rws_read_sqlite_init}
\title{Read Initialization Data table from SQLite Database}
\usage{
rws_read_init(conn)
}
\arguments{
\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}
}
\value{
A data frame of the init table
}
\description{
The table is created if it doesn't exist.
}
\examples{
conn <- rws_connect()
rws_read_init(conn)
rws_write(rws_data, exists = FALSE, conn = conn)
rws_read_init(conn)
rws_disconnect(conn)
}
