% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{rws_read_log}
\alias{rws_read_log}
\alias{rws_read_sqlite_log}
\title{Read Log Data Table from a SQLite Database}
\usage{
rws_read_log(conn)
}
\arguments{
\item{conn}{A \linkS4class{SQLiteConnection} to a database.}
}
\value{
A data frame of the log table
}
\description{
The table is created if it doesn't exist.
}
\examples{
conn <- rws_connect()
rws_read_log(conn)
rws_write(rws_data, exists = FALSE, conn = conn)
\dontrun{
rws_read_log(conn)
}
rws_disconnect(conn)
}
