% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{rws_read_sqlite.character}
\alias{rws_read_sqlite.character}
\title{Read from a SQLite Database}
\usage{
\method{rws_read_sqlite}{character}(x, conn = getOption("rws.conn",
  NULL), meta = TRUE, ...)
}
\arguments{
\item{x}{The object to write.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{...}{Not used.}
}
\value{
A data frame (tibble if tibble package is installed) of the table.
}
\description{
Read from a SQLite Database
}
\seealso{
Other rws_read_sqlite: \code{\link{rws_read_sqlite.SQLiteConnection}},
  \code{\link{rws_read_sqlite}}
}
\concept{rws_read_sqlite}
