% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{rws_read_sqlite_log}
\alias{rws_read_sqlite_log}
\title{Read Log Data Table from SQLite Database}
\usage{
rws_read_sqlite_log(conn = getOption("rws.conn", NULL))
}
\arguments{
\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}
}
\value{
A data frame of the log table
}
\description{
The table is created if it doesn't exist.
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite())
rws_read_sqlite_meta(con)
DBI::dbDisconnect(con)
}
