% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{rws_read_sqlite_table}
\alias{rws_read_sqlite_table}
\title{Read A Tables from a SQLite Database}
\usage{
rws_read_sqlite_table(x, meta = TRUE, conn)
}
\arguments{
\item{x}{A string of the table name.}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}
}
\value{
A data frame of the table.
}
\description{
Read A Tables from a SQLite Database
}
