% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{rws_write_sqlite.environment}
\alias{rws_write_sqlite.environment}
\title{Write the Data Frames in an Environment to a SQLite Database}
\usage{
\method{rws_write_sqlite}{environment}(x, exists = TRUE,
  delete = FALSE, commit = TRUE, meta = TRUE, log = TRUE,
  strict = TRUE, x_name = substitute(x),
  silent = getOption("rws.silent", FALSE), conn, complete = TRUE, ...)
}
\arguments{
\item{x}{An environment.}

\item{exists}{A flag specifying whether the table(s) must already exist.}

\item{delete}{A flag specifying whether to delete existing rows before 
inserting data.}

\item{commit}{A flag specifying whether to commit the operations 
(calling with commit = FALSE can be useful for checking data).}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{log}{A flag specifying whether to log the table operations.}

\item{strict}{A flag specifying whether to error if x has extraneous columns or if exists = TRUE extraneous data frames.}

\item{x_name}{A string of the name of the object.}

\item{silent}{A flag specifying whether to suppress messages and warnings.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}

\item{complete}{A flag specifying whether all the tables in the data base must be represented.}

\item{...}{Not used.}
}
\description{
Write the Data Frames in an Environment to a SQLite Database
}
\seealso{
Other rws_write_sqlite: \code{\link{rws_write_sqlite.data.frame}},
  \code{\link{rws_write_sqlite.list}},
  \code{\link{rws_write_sqlite}}
}
\concept{rws_write_sqlite}
