% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{rws_write_sqlite.list}
\alias{rws_write_sqlite.list}
\title{Write a Named List of Data Frames to a SQLite Database}
\usage{
\method{rws_write_sqlite}{list}(x, exists = getOption("rws.exists", NA),
  delete = FALSE, commit = TRUE, meta = TRUE, log = TRUE,
  conn = getOption("rws.conn", NULL), x_name = substitute(x),
  silent = getOption("rws.silent", FALSE), complete = FALSE, ...)
}
\arguments{
\item{x}{A named list of data frames.}

\item{exists}{A flag specifying whether the table must already exist.}

\item{delete}{A flag specifying whether to delete existing rows before 
inserting data.}

\item{commit}{A flag specifying whether to commit the operations 
(calling with commit = FALSE can be useful for checking data).}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{log}{A flag specifying whether to log the operations alterations.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}

\item{x_name}{A string of the name of the object.}

\item{silent}{A flag specifying whether suppress messages and warnings.}

\item{complete}{A flag specifying whether all the tables in the data base must be represented.
If complete = TRUE and exists = TRUE then extra tables are ignored (with a warning).}

\item{...}{Not used.}
}
\description{
Write a Named List of Data Frames to a SQLite Database
}
\seealso{
Other rws_write_sqlite: \code{\link{rws_write_sqlite.data.frame}},
  \code{\link{rws_write_sqlite.environment}},
  \code{\link{rws_write_sqlite}}
}
\concept{rws_write_sqlite}
