% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite-name.R
\name{sqlite_name}
\alias{sqlite_name}
\title{SQLite Name}
\description{
A SQLite name is a vector of names for SQLite objects that inherits from character.
It is useful for comparing SQLite names which are case insensitive except when 
they are quoted using \code{"name"}, \code{[name]} \code{`name`}.
When making a comparison, SQLite names are converted to upper case 
unless they are quoted using one of the three methods 
in which case the quotes are removed.
}
\examples{
x <- 1
}
