% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_sqlite.data.frame}
\alias{write_sqlite.data.frame}
\title{Write a data frame to a SQLite Database}
\usage{
\method{write_sqlite}{data.frame}(x,
  conn = getOption("readwritesqlite.conn", NULL), delete = FALSE,
  commit = TRUE, meta = TRUE, log = TRUE,
  table_name = substitute(x), ...)
}
\arguments{
\item{x}{A data frame.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}

\item{delete}{A flag specifying whether to delete existing rows before 
inserting data.}

\item{commit}{A flag specifying whether to commit the data alterations 
(useful for checking data).}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{log}{A flag specifying whether to log data alterations.}

\item{table_name}{A string of the (case insensitive) table name.}

\item{...}{Not used.}
}
\description{
Write a data frame to a SQLite Database
}
\seealso{
Other write_sqlite: \code{\link{read_sqlite}},
  \code{\link{write_sqlite.list}},
  \code{\link{write_sqlite}}
}
\concept{write_sqlite}
