% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_sqlite.list}
\alias{write_sqlite.list}
\title{Write a list of data frames to a SQLite Database}
\usage{
\method{write_sqlite}{list}(x, conn = getOption("readwritesqlite.conn",
  NULL), delete = FALSE, commit = TRUE, meta = TRUE, log = TRUE,
  ...)
}
\arguments{
\item{x}{A named list of data frames.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}

\item{delete}{A flag specifying whether to delete existing rows before 
inserting data.}

\item{commit}{A flag specifying whether to commit the data alterations 
(useful for checking data).}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{log}{A flag specifying whether to log data alterations.}

\item{...}{Not used.}
}
\description{
Write a list of data frames to a SQLite Database
}
\seealso{
Other write_sqlite: \code{\link{read_sqlite}},
  \code{\link{write_sqlite.data.frame}},
  \code{\link{write_sqlite}}
}
\concept{write_sqlite}
