% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_sqlite}
\alias{write_sqlite}
\title{Write to a SQLite Database}
\usage{
write_sqlite(x, conn = getOption("readwritesqlite.conn", NULL),
  delete = FALSE, commit = TRUE, meta = TRUE, log = TRUE, ...)
}
\arguments{
\item{x}{The object to write.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}

\item{delete}{A flag specifying whether to delete existing rows before 
inserting data.}

\item{commit}{A flag specifying whether to commit the data alterations 
(useful for checking data).}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{log}{A flag specifying whether to log data alterations.}

\item{...}{Not used.}
}
\value{
An invisible character vector of the names of the tables that were altered.
}
\description{
write_sqlite() does not create any tables other than 
readwritesqlite_log and readwritesqlite_meta for logging data alterations
and preserving meta data.
}
\seealso{
Other write_sqlite: \code{\link{read_sqlite}},
  \code{\link{write_sqlite.data.frame}},
  \code{\link{write_sqlite.list}}
}
\concept{write_sqlite}
