% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite-name.R
\name{as.sqlite_name}
\alias{as.sqlite_name}
\title{SQLite Name}
\usage{
as.sqlite_name(x, ...)
}
\arguments{
\item{x}{An object to coerce.}

\item{...}{Unused}
}
\description{
A SQLite name is a vector of names for SQLite objects that inherits from character.
It is useful for comparing SQLite names which are case insensitive 
except when they are quoted using \code{"name"}, \code{[name]} \code{`name`}.
}
