% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data.R
\name{dbCheckDataSQLite}
\alias{dbCheckDataSQLite}
\title{Check and possibly convert a Data Frame}
\usage{
dbCheckDataSQLite(conn, name, value, order = FALSE, exclusive = FALSE,
  convert = FALSE)
}
\arguments{
\item{conn}{A \code{\linkS4class{SQLiteConnection}} object, produced by
[DBI::dbConnect()].}

\item{name}{A string of the (case insensitive) table name.}

\item{value}{A data frame of the data to write.}

\item{order}{A flag indicating whether the columns have to occur in the same order as values.}

\item{exclusive}{A flag indicating whether other columns are not permitted.}

\item{convert}{A flag specifying whether to convert data columns.}
}
\value{
The data frame with the required and converted columns in the correct order.
}
\description{
Factors are converted to character,
raw to character and character to UTF-8 encoding.
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite())
DBI::dbDisconnect(con)
}
