% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{dbReadTablesSQLite}
\alias{dbReadTablesSQLite}
\title{Read Tables from a SQLite Database}
\usage{
dbReadTablesSQLite(conn = getOption("dbWriteSQLite.conn", NULL),
  meta = TRUE)
}
\arguments{
\item{conn}{A \code{\linkS4class{SQLiteConnection}} object.}

\item{meta}{A flag specifying whether to preserve meta data.}
}
\value{
A list of the tables.
}
\description{
Read Tables from a SQLite Database
}
\seealso{
Other dbReadTableSQLite: \code{\link{dbReadTableSQLite}}
}
\concept{dbReadTableSQLite}
