% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-table.R
\name{dbWriteTableSQLite}
\alias{dbWriteTableSQLite}
\title{Write a local data frame or file to the database}
\usage{
dbWriteTableSQLite(conn, name, value)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[=dbConnect]{dbConnect()}}.}

\item{name}{A character string specifying the unquoted DBMS table name,
or the result of a call to \code{\link[=dbQuoteIdentifier]{dbQuoteIdentifier()}}.}

\item{value}{a \link{data.frame} (or coercible to data.frame).}
}
\description{
Functions for writing data frames or delimiter-separated files
to database tables.
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite())
DBI::dbDisconnect(con)
}
