% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-table.R
\name{dbWriteTableSQLite}
\alias{dbWriteTableSQLite}
\title{Write a data frame to a SQLite Database}
\usage{
dbWriteTableSQLite(conn, name, value, commit = TRUE, meta = TRUE,
  log = TRUE)
}
\arguments{
\item{conn}{A \code{\linkS4class{SQLiteConnection}} object, produced by
[DBI::dbConnect()].}

\item{name}{A string of the (case insensitive) table name.}

\item{value}{A data frame of the data to write.}

\item{commit}{A flag specifying whether to commit the changes.}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{log}{A flag specifying whether to log change.}
}
\description{
The table must exist.
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite())
DBI::dbDisconnect(con)
}
