% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite-name.R
\name{is.sqlite_name}
\alias{is.sqlite_name}
\title{Is SQLite Name}
\usage{
is.sqlite_name(x)
}
\arguments{
\item{x}{An R object to test.}
}
\value{
A flag specifying whether an R object is a sqlite_name.
}
\description{
Tests whether an R object inherits from class \code{sqlite_name}.
}
