% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwritesqlite-package.R
\docType{package}
\name{readwritesqlite-package}
\alias{readwritesqlite}
\alias{readwritesqlite-package}
\title{readwritesqlite: Enhanced Reading and Writing for SQLite Databases}
\description{
Preserves time zones (for POSIXct columns), 
  projections (for sfc columns), units (for unit columns), 
  levels (for factors and ordered factors) and classes for 
  logical, Date and hms columns.
  It also logs changes to tables and provides more informative error messages.
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

}
\keyword{internal}
