% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{rws_describe_meta}
\alias{rws_describe_meta}
\alias{rws_describe_sqlite_meta}
\title{Add Descriptions to SQL Meta Data Table}
\usage{
rws_describe_meta(x, ..., conn)
}
\arguments{
\item{x}{An object specifying the table(s) to read.}

\item{...}{Not used.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}
}
\value{
xx
}
\description{
Add Descriptions to SQL Meta Data Table
}
\concept{rws_describe_sqlite_meta}
