% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{rws_describe_meta}
\alias{rws_describe_meta}
\title{Add Descriptions to SQL Meta Data Table}
\usage{
rws_describe_meta(x, ..., conn)
}
\arguments{
\item{x}{An object specifying the descriptions.}

\item{...}{Not used.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}
}
\value{
An invisible copy of the updated meta table.
}
\description{
Add Descriptions to SQL Meta Data Table
}
\seealso{
Other rws_describe_meta: \code{\link{rws_describe_meta.character}}
}
\concept{rws_describe_meta}
