% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{rws_disconnect}
\alias{rws_disconnect}
\alias{rws_close_connection}
\title{Close SQLite Database Connection}
\usage{
rws_disconnect(conn)
}
\arguments{
\item{conn}{An \code{RSQLite::SQLiteConnection()}.}
}
\description{
Closes a \linkS4class{SQLiteConnection} to a SQLite database.
}
\examples{
conn <- rws_connect()
rws_disconnect(conn)
print(conn)
}
\seealso{
\code{\link[=rws_connect]{rws_connect()}}
}
