% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{rws_read_sqlite}
\alias{rws_read_sqlite}
\title{Read from a SQLite Database}
\usage{
rws_read_sqlite(x, ...)
}
\arguments{
\item{x}{An object specifying the table(s) to read.}

\item{...}{Not used.}
}
\value{
A named list of data frames.
}
\description{
Read from a SQLite Database
}
\seealso{
Other rws_read_sqlite: \code{\link{rws_describe_sqlite_meta.character}},
  \code{\link{rws_describe_sqlite_meta.data.frame}},
  \code{\link{rws_read_sqlite.SQLiteConnection}},
  \code{\link{rws_read_sqlite.character}}
}
\concept{rws_read_sqlite}
