% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{rws_read_sqlite.SQLiteConnection}
\alias{rws_read_sqlite.SQLiteConnection}
\title{Read from a SQLite Database}
\usage{
\method{rws_read_sqlite}{SQLiteConnection}(x = getOption("rws.conn",
  NULL), meta = TRUE, ...)
}
\arguments{
\item{x}{The object to write.}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{...}{Not used.}
}
\value{
A named list of the data tables.
}
\description{
Read from a SQLite Database
}
\concept{dbReadTableSQLite}
