% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{rws_read_sqlite_meta}
\alias{rws_read_sqlite_meta}
\title{Read Meta Data table from SQLite Database}
\usage{
rws_read_sqlite_meta(conn = getOption("rws.conn", NULL))
}
\arguments{
\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}
}
\value{
A data frame of the meta table
}
\description{
The table is created if it doesn't exist.
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite())
rws_read_sqlite_meta(con)
DBI::dbDisconnect(con)
}
