% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{rws_write_sqlite.list}
\alias{rws_write_sqlite.list}
\title{Write a list of data frames to a SQLite Database}
\usage{
\method{rws_write_sqlite}{list}(x, conn = getOption("rws.conn", NULL),
  exists = TRUE, delete = FALSE, commit = TRUE, meta = TRUE,
  log = TRUE, ...)
}
\arguments{
\item{x}{A named list of data frames.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}

\item{exists}{A flag specifying whether the table must already exist.}

\item{delete}{A flag specifying whether to delete existing rows before 
inserting data.}

\item{commit}{A flag specifying whether to commit the operations 
(calling with commit = FALSE can be useful for checking data).}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{log}{A flag specifying whether to log the operations alterations.}

\item{...}{Not used.}
}
\description{
Write a list of data frames to a SQLite Database
}
\seealso{
Other rws_write_sqlite: \code{\link{rws_write_sqlite.data.frame}},
  \code{\link{rws_write_sqlite}}
}
\concept{rws_write_sqlite}
