% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{rws_write_sqlite}
\alias{rws_write_sqlite}
\title{Write to a SQLite Database}
\usage{
rws_write_sqlite(x, exists = getOption("rws.exists", NA),
  delete = FALSE, commit = TRUE, meta = TRUE, log = TRUE,
  conn = getOption("rws.conn", NULL), ...)
}
\arguments{
\item{x}{The object to write.}

\item{exists}{A flag specifying whether the table must already exist.}

\item{delete}{A flag specifying whether to delete existing rows before 
inserting data.}

\item{commit}{A flag specifying whether to commit the operations 
(calling with commit = FALSE can be useful for checking data).}

\item{meta}{A flag specifying whether to preserve meta data.}

\item{log}{A flag specifying whether to log the operations alterations.}

\item{conn}{A \code{\linkS4class{SQLiteConnection}} to a database.}

\item{...}{Not used.}
}
\value{
An invisible character vector of the name(s) of the table(s).
}
\description{
Write to a SQLite Database
}
\seealso{
Other rws_write_sqlite: \code{\link{rws_write_sqlite.data.frame}},
  \code{\link{rws_write_sqlite.list}}
}
\concept{rws_write_sqlite}
